@@ -3,25 +3,27 @@
   "version": "1.0.0",
   "description": "Rapidfire Express Session Example",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "start": "node src/index",
+    "dev": "npm start",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
   "license": "BEERWARE",
   "dependencies": {
-    "@luasenvy/rapidfire": "^0.1.8",
-    "express-session": "^1.17.1"
+    "@luasenvy/rapidfire": "^0.3.7",
+    "express-session": "^1.17.3"
   },
   "devDependencies": {
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.19.0",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "husky": "^4.3.8",
+    "@babel/eslint-parser": "^7.17.0",
+    "eslint": "^8.15.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "husky": "^8.0.1",
     "imagemin-lint-staged": "^0.4.0",
-    "lint-staged": "^10.5.4",
-    "prettier": "^2.2.1"
+    "lint-staged": "^12.4.1",
+    "prettier": "^2.6.2"
   }
 }
